fix: renaming schemes -> schemas
