+++ services/metrics/app/api/routers/agent.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, HTTPException, Depends
 
-from app.api.schemes import AgentResponse
+from app.api.schemas import AgentResponse
 from app.api.deps import get_agent_metrics_manager, AgentsResponseManager
 from app.logs import get_logger
