+++ services/metrics/app/api/routers/info.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, HTTPException, status
 from fastapi.responses import JSONResponse
 
 from app.core.health import check_health_all
-from app.api.schemes import HealthResponse
+from app.api.schemas import HealthResponse
 
 routers = APIRouter(
     prefix="/info",
